feat: MCP Token Compression & Context Bounds Integration

+++ src/indexer/extractor.rs
@@ -314,7 +314,7 @@ impl<'a> EntityExtractor<'a> {
         }
 
         for i in 0..node.child_count() {
-            if let Some(child) = node.child(i) {
+            if let Some(child) = node.child(i as u32) {
                 let current_parent = if matches!(
                     node_type,
                     "function_declaration"
